+++ Models/IBusinessLogic.cs
@@ -38,7 +38,7 @@ public interface IBusinessLogic {
     /// <returns>
     /// returns the observable collection if lookup is successful otherwise null
     /// </returns>
-    Task<ObservableCollection<Question>?> GetQuestions(long id);
+    Task<ObservableCollection<Question>?> GetQuestions(int id);
 
     /// <summary>
     /// Adds a question to the questions table
